requirements.txt
google-genai
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-dotenv
python-multipart